rename util appropriate- egglog_to_latex
