+++ sealir-tutorials/egglog_to_latex.py
